added support for out-of-corpus documents in doc2vec similarity
Although gensim's documentation makes it sound like most_similar() is
not possible for out-of-corpus documents with current functionality, it
actually can be done. First, the out-of-corpus document's vector is
inferred with infer_vector(). Then we call most_common() with the
inferred vector (weighted 1.0 for the under-the-hood averaging that
gensim does over all the positive examples given to most_similar())
instead of the doc ID.